+++ src/main/kotlin/net/pwall/log/LoggerUtil.kt
@@ -29,30 +29,107 @@ import kotlin.reflect.KClass
 import java.time.Clock
 
 /**
- * Get a [Logger] for a Kotlin [KClass], using the optional [Level] and [Clock]. (Kotlin equivalent of same function
+ * Get a [Logger] from the default [LoggerFactory], using the name of the calling class and the default level and clock.
+ */
+fun getLogger(): Logger = Log.getLogger()
+
+/**
+ * Get a [Logger] from the default [LoggerFactory], using the name of the calling class, the specified level and the
+ * default clock.
+ */
+fun getLogger(level: Level): Logger = Log.getLogger(level)
+
+/**
+ * Get a [Logger] from the default [LoggerFactory], using the name of the calling class, the default level and the
+ * specified clock.
+ */
+fun getLogger(clock: Clock): Logger = Log.getLogger(clock)
+
+/**
+ * Get a [Logger] from the default [LoggerFactory], using the name of the calling class and the specified level and
+ * clock.
+ */
+fun getLogger(level: Level, clock: Clock): Logger = Log.getLogger(level, clock)
+
+/**
+ * Get a [Logger] from the default [LoggerFactory], using the specified name and the default level and clock.
+ */
+fun getLogger(name: String): Logger = Log.getLogger(name)
+
+/**
+ * Get a [Logger] from the default [LoggerFactory], using the specified name and level and the default clock.
+ */
+fun getLogger(name: String, level: Level): Logger = Log.getLogger(name, level)
+
+/**
+ * Get a [Logger] from the default [LoggerFactory], using the specified name and clock and the default level.
+ */
+fun getLogger(name: String, clock: Clock): Logger = Log.getLogger(name, clock)
+
+/**
+ * Get a [Logger] from the default [LoggerFactory], using the specified name, level and clock.
+ */
+fun getLogger(name: String, level: Level, clock: Clock): Logger = Log.getLogger(name, level, clock)
+
+/**
+ * Get a [Logger] from the default [LoggerFactory], using the specified Java [Class] name and the default level and
+ * clock.
+ */
+fun getLogger(javaClass: Class<*>): Logger = Log.getLogger(javaClass.name)
+
+/**
+ * Get a [Logger] from the default [LoggerFactory], using the specified Java [Class] name and level and the default
+ * clock.
+ */
+fun getLogger(javaClass: Class<*>, level: Level): Logger = Log.getLogger(javaClass.name, level)
+
+/**
+ * Get a [Logger] from the default [LoggerFactory], using the specified Java [Class] name and clock and the default
+ * level.
+ */
+fun getLogger(javaClass: Class<*>, clock: Clock): Logger = Log.getLogger(javaClass.name, clock)
+
+/**
+ * Get a [Logger] from the default [LoggerFactory], using the specified Java [Class] name, level and clock.
+ */
+fun getLogger(javaClass: Class<*>, level: Level, clock: Clock): Logger = Log.getLogger(javaClass.name, level, clock)
+
+/**
+ * Get a [Logger] from the default [LoggerFactory], using the specified [KClass] name and the default level and clock.
+ */
+fun getLogger(kClass: KClass<*>): Logger = Log.getLogger(kClass.qualifiedName)
+
+/**
+ * Get a [Logger] from the default [LoggerFactory], using the specified [KClass] name and level and the default clock.
+ */
+fun getLogger(kClass: KClass<*>, level: Level): Logger = Log.getLogger(kClass.qualifiedName, level)
+
+/**
+ * Get a [Logger] from the default [LoggerFactory], using the specified [KClass] name and clock and the default level.
+ */
+fun getLogger(kClass: KClass<*>, clock: Clock): Logger = Log.getLogger(kClass.qualifiedName, clock)
+
+/**
+ * Get a [Logger] from the default [LoggerFactory], using the specified [KClass] name, level and clock.
+ */
+fun getLogger(kClass: KClass<*>, level: Level, clock: Clock): Logger = Log.getLogger(kClass.qualifiedName, level, clock)
+
+/**
+ * Get a [Logger] for a Kotlin [KClass], using the optional [Level] and [Clock].  (Kotlin equivalent of same function
  * taking a Java [Class]).
  *
  * @receiver            [LoggerFactory]
- * @param   kotlinClass a Kotlin [KClass]
+ * @param   kClass      a Kotlin [KClass]
  * @param   level       the [Level]
  * @param   clock       the [Clock]
  * @param   <L>         the [Logger] type
  * @return              a [Logger]
  */
-fun <L: Logger> LoggerFactory<L>.getLogger(
-    kotlinClass: KClass<*>,
+fun <L : Logger> LoggerFactory<L>.getLogger(
+    kClass: KClass<*>,
     level: Level = defaultLevel,
     clock: Clock = defaultClock,
-): L = getLogger(kotlinClass.qualifiedName, level, clock)
-
-/**
- * Output an error message supplied by a lambda, along with a [Throwable].  This duplicates a function in the [Logger]
- * class, but with the parameters reversed to allow the use of Kotlin lambda syntax.
- *
- * @param   throwable           the [Throwable]
- * @param   messageSupplier     the message supplier lambda
- */
-fun Logger.error(throwable: Throwable, messageSupplier: () -> Any) = error(messageSupplier, throwable)
+): L = getLogger(kClass.qualifiedName, level, clock)
 
 /**
  * Test whether a [LogItem] represents a message with level `TRACE` and the specified text.
@@ -61,7 +138,7 @@ fun Logger.error(throwable: Throwable, messageSupplier: () -> Any) = error(messa
  * @param   text    the text to be compared
  * @return          `true` if the [LogItem] matches
  */
-infix fun LogItem.isTrace(text: String): Boolean = level == Level.TRACE && this.text == text
+infix fun LogItem.isTrace(text: String): Boolean = level == Level.TRACE && message == text
 
 /**
  * Test whether a [LogItem] represents a message with level `TRACE` and the specified text.
@@ -70,7 +147,7 @@ infix fun LogItem.isTrace(text: String): Boolean = level == Level.TRACE && this.
  * @param   regex   the [Regex]
  * @return          `true` if the [LogItem] matches
  */
-infix fun LogItem.isTrace(regex: Regex): Boolean = level == Level.TRACE && regex.containsMatchIn(text)
+infix fun LogItem.isTrace(regex: Regex): Boolean = level == Level.TRACE && regex.containsMatchIn(message.toString())
 
 /**
  * Test whether a [LogItem] represents a message with level `DEBUG` and the specified text.
@@ -79,7 +156,7 @@ infix fun LogItem.isTrace(regex: Regex): Boolean = level == Level.TRACE && regex
  * @param   text    the text to be compared
  * @return          `true` if the [LogItem] matches
  */
-infix fun LogItem.isDebug(text: String): Boolean = level == Level.DEBUG && this.text == text
+infix fun LogItem.isDebug(text: String): Boolean = level == Level.DEBUG && message == text
 
 /**
  * Test whether a [LogItem] represents a message with level `DEBUG` and the specified text.
@@ -88,7 +165,7 @@ infix fun LogItem.isDebug(text: String): Boolean = level == Level.DEBUG && this.
  * @param   regex   the [Regex]
  * @return          `true` if the [LogItem] matches
  */
-infix fun LogItem.isDebug(regex: Regex): Boolean = level == Level.DEBUG && regex.containsMatchIn(text)
+infix fun LogItem.isDebug(regex: Regex): Boolean = level == Level.DEBUG && regex.containsMatchIn(message.toString())
 
 /**
  * Test whether a [LogItem] represents a message with level `INFO` and the specified text.
@@ -97,7 +174,7 @@ infix fun LogItem.isDebug(regex: Regex): Boolean = level == Level.DEBUG && regex
  * @param   text    the text to be compared
  * @return          `true` if the [LogItem] matches
  */
-infix fun LogItem.isInfo(text: String): Boolean = level == Level.INFO && this.text == text
+infix fun LogItem.isInfo(text: String): Boolean = level == Level.INFO && message == text
 
 /**
  * Test whether a [LogItem] represents a message with level `INFO` and text that matches the given [Regex].
@@ -106,7 +183,7 @@ infix fun LogItem.isInfo(text: String): Boolean = level == Level.INFO && this.te
  * @param   regex   the [Regex]
  * @return          `true` if the [LogItem] matches
  */
-infix fun LogItem.isInfo(regex: Regex): Boolean = level == Level.INFO && regex.containsMatchIn(text)
+infix fun LogItem.isInfo(regex: Regex): Boolean = level == Level.INFO && regex.containsMatchIn(message.toString())
 
 /**
  * Test whether a [LogItem] represents a message with level `WARN` and the specified text.
@@ -115,7 +192,7 @@ infix fun LogItem.isInfo(regex: Regex): Boolean = level == Level.INFO && regex.c
  * @param   text    the text to be compared
  * @return          `true` if the [LogItem] matches
  */
-infix fun LogItem.isWarning(text: String): Boolean = level == Level.WARN && this.text == text
+infix fun LogItem.isWarning(text: String): Boolean = level == Level.WARN && message == text
 
 /**
  * Test whether a [LogItem] represents a message with level `WARN` and text that matches the given [Regex].
@@ -124,7 +201,7 @@ infix fun LogItem.isWarning(text: String): Boolean = level == Level.WARN && this
  * @param   regex   the [Regex]
  * @return          `true` if the [LogItem] matches
  */
-infix fun LogItem.isWarning(regex: Regex): Boolean = level == Level.WARN && regex.containsMatchIn(text)
+infix fun LogItem.isWarning(regex: Regex): Boolean = level == Level.WARN && regex.containsMatchIn(message.toString())
 
 /**
  * Test whether a [LogItem] represents a message with level `ERROR` and the specified text.
@@ -133,7 +210,7 @@ infix fun LogItem.isWarning(regex: Regex): Boolean = level == Level.WARN && rege
  * @param   text    the text to be compared
  * @return          `true` if the [LogItem] matches
  */
-infix fun LogItem.isError(text: String): Boolean = level == Level.ERROR && this.text == text
+infix fun LogItem.isError(text: String): Boolean = level == Level.ERROR && message == text
 
 /**
  * Test whether a [LogItem] represents a message with level `ERROR` and text that matches the given [Regex].
@@ -142,4 +219,4 @@ infix fun LogItem.isError(text: String): Boolean = level == Level.ERROR && this.
  * @param   regex   the [Regex]
  * @return          `true` if the [LogItem] matches
  */
-infix fun LogItem.isError(regex: Regex): Boolean = level == Level.ERROR && regex.containsMatchIn(text)
+infix fun LogItem.isError(regex: Regex): Boolean = level == Level.ERROR && regex.containsMatchIn(message.toString())